Refactor ModelsContent to replace RightBar with ModelsRightBar for improved session management
